Fixes #13 Users can create a diary entry Moved test to src folder
